add usePrefetchImages
